dataplane: unhang and repair the privilege-gated cells
Running the 42 memlock-gated guards under CAP_BPF for the first time
(#9337) left ./pkg/dataplane/... with eight red cells and one VOID: a
15-minute package timeout. None was a second production defect; all were
fixture or expectation defects that could not be seen where the suite
normally runs. Leaving them red would teach people to ignore the leg,
which is the outcome pkg/memlockcensus set out to avoid.

THE HANG, first, because its blast radius exceeded its own cell. The
goroutine dump named it: TestManager_ArmedGate_DetachRetainedClaims
waited forever on <-probeArrived while DetachXDP sat in
sync.Mutex.Lock inside xdpLinkFor. The detach DID block on m.mu -- but
at a surface carrying no muAcquireProbeHook pre-lock fire, so the
arrival proof never fired. It was not always so: before 01409c1f
(2026-08-22, "guard the link maps behind m.mu and hand out snapshots")
DetachXDP read m.xdpLinks unlocked and its first contended acquisition
WAS the probed lookupMapLocked. That commit moved a plain acquisition
ahead of the probed one and silently converted a valid arrival proof
into an unbounded deadlock. The cell needs a real BPF map, so it skips
unprivileged and nothing could see it.

xdpLinkFor now fires the probe (the armedGateDirectLockSet attribution
legs still pass, so the pin did not drift), and the wait is BOUNDED with
a named failure. The budget a hang burns belongs to every later cell:
two further defects below were only reachable once this stopped eating
the package timeout. The hook's doc comment records that the probed set
is a claim about the CALLERS, not about those functions.

The rest:

  * Four cells reach applyHelperStatusLocked, which since #6994 also
    re-syncs the ingress/local/interface-NAT classifier maps; the
    fixtures load only ctrl+bindings, so it failed with
    "userspace_ingress_ifaces map not loaded" before reaching anything
    they assert. They take the #7468 syncClassifierMapsHook seam.

  * TestXSKLivenessFailureRestoresUserspaceShimEntry additionally
    carried an expectation hollowed by #6429: an expired probe with no
    RX now resolves three ways, and shouldExtendXSKLivenessIdleLocked
    EXTENDS whenever the node has no active data RG -- which an empty
    haGroups map means. The cell measured the extension path while
    asserting the failure path's outcome. It now seeds an active data RG.

  * TestSyncedImportRefusalRollsBackWithoutGatingTakeover6785 used
    t.TempDir(), whose directory name embeds the full test and sub-test
    names; the resulting socket path exceeds the 108-byte AF_UNIX
    sun_path limit and binds with EINVAL, which reads like a dataplane
    defect. A short MkdirTemp prefix bounds it under any rename.

  * The two TestUserspaceXDP* cells seeded ingress ifindex 0.
    BPF_PROG_TEST_RUN does not present 0: bpf_prog_test_run_xdp
    registers the netns loopback device as the run's RX queue, so
    ctx->ingress_ifindex is lo's and the binding coordinate is
    ifindex * BINDING_QUEUES_PER_IFACE + queue. Seeded at 0 the shim
    never recognised the ingress interface and XDP_PASSed with no
    degraded reason recorded at all -- hence "action = 2, want 1" rather
    than a wrong reason code. The NDP cell also never seeded
    userspace_heartbeat, so heartbeat_stale fired before early_filter;
    it is seeded on CLOCK_MONOTONIC, the clock bpf_ktime_get_ns reads,
    because a CLOCK_BOOTTIME value on a suspended host lands in the
    shim's future and its own guard reads that as stale.

  * Revealed by unhanging: Test_clear_global_counters_read_equation_
    zero_5098 passed an UNTYPED constant as a BPF map key, which boxes
    as int and cilium/ebpf refuses.

  * Revealed by unhanging: TestReconcileProxyARP_V6InstallsPneighLive
    read a channel that cannot carry its answer. A plain RTM_GETNEIGH
    dump has ndm_flags == 0 and the kernel answers it by walking the
    NEIGHBOUR tables; it walks PNEIGH only when the request sets
    NTF_PROXY. NeighList was therefore RED against correct code as
    surely as against the pre-#2197 code it names -- it asserted
    nothing. It now uses NeighProxyList, and installs a pneigh entry it
    owns as a POSITIVE CONTROL: a prober that cannot see proxy entries
    and a reconcile that installed nothing report the same thing, and
    that reading is exactly the defect. Both directions were
    mutation-checked -- reverting the channel fires the control message,
    severing the production v6 install fires the other.

assertUserspaceXDPDegradedPathStat now names the reasons that DID fire.
The shim's degraded gates are an ordered chain, so "stat X = 0" is
almost always "an earlier gate diverted the packet", which the counter
set states directly; diagnosing it from the bare message cost a full
read of userspace-xdp/src/lib.rs.

Validation: sudo go test ./pkg/dataplane/... is green and pkg/dataplane
completes in ~31s instead of timing out at 900s. go test ./... green.

Refs #9337

+++ pkg/dataplane/maps_counters_clear_5098_test.go
@@ -37,7 +37,14 @@ func Test_clear_global_counters_read_equation_zero_5098(t *testing.T) {
 	numCPUs := ebpf.MustPossibleCPU()
 	arr := make([]uint64, numCPUs)
 	arr[0] = 7
-	if err := gc.Update(idx, arr, ebpf.UpdateAny); err != nil {
+	// #9337: uint32(idx), not idx. `const idx = GlobalCtrRxPackets` is an
+	// UNTYPED constant, so passing it to Update's interface{} key boxes it as
+	// `int` and cilium/ebpf refuses with "some values are not fixed-sized in
+	// type int" — the map key is 4 bytes. The sibling calls below take a typed
+	// parameter and convert implicitly, which is why only this one failed. The
+	// cell needs a real BPF map, so it skips unprivileged and this never
+	// surfaced under `make test-go`.
+	if err := gc.Update(uint32(idx), arr, ebpf.UpdateAny); err != nil {
 		t.Fatalf("seed global_counters array: %v", err)
 	}
 	if err := m.IncrementGlobalCounter(idx, 100); err != nil {
